requirements.txt
agora-token-builder==1.0.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
constantly==15.1.0
contourpy==1.0.7
cryptography==3.4.7
cycler==0.11.0
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.7
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
filelock==3.8.0
fonttools==4.39.2
hyperlink==21.0.0
idna==3.2
imageio==2.26.0
importlib-resources==5.12.0
incremental==21.3.0
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
oauthlib==3.1.1
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.0.1
platformdirs==2.5.2
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.4.0
pymongo==3.12.1
PyMuPDF==1.21.1
pyOpenSSL==20.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.4.1
requests==2.26.0
requests-oauthlib==1.3.0
scikit-image==0.20.0
scipy==1.9.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
sympy==1.11.1
tifffile==2023.3.15
torch==2.0.0
torchvision==0.15.1
Twisted==22.4.0
twisted-iocpsupport==1.0.4
txaio==21.2.1
typing-extensions==3.10.0.0
tzdata==2022.1
urllib3==1.26.6
virtualenv==20.16.6
zipp==3.15.0
zope.interface==5.4.0
